backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tensorflow>=2.16.0
pillow>=10.0.0
python-multipart>=0.0.9
numpy>=1.26.0
